requirements.txt
beautifulsoup4==4.11.1
Flask==1.0.2
Flask_APScheduler==1.12.4
Pillow==9.3.0
PyYAML==6.0
